+++ src/client/java/mchorse/bbs_mod/ui/framework/UIScreen.java
@@ -146,7 +146,7 @@ public class UIScreen extends Screen implements IFileDropListener
     @Override
     public boolean mouseScrolled(double mouseX, double mouseY, double horizontalAmount, double verticalAmount)
     {
-        return this.menu.mouseScrolled((int) mouseX, (int) mouseY, (int) verticalAmount);
+        return this.menu.mouseScrolled((int) mouseX, (int) mouseY, verticalAmount);
     }
 
     @Override
